+++ src/Mobile.BuildTools.Common/Mobile.BuildTools.Common.targets
