+++ src/Mobile.BuildTools.Images/Mobile.BuildTools.Images.csproj
@@ -15,7 +15,6 @@
     <None Include="*.targets;*.props" Pack="true" PackFolder="$(PackFolder)" />
     <None Update="buildTransitive\*.targets;buildTransitive\*.props" Pack="true" PackFolder="buildTransitive" />
     <Compile Include="..\Mobile.BuildTools.Common\Build\BuildExtensions.cs" LinkBase="Build" />
-    <Compile Include="..\Mobile.BuildTools\Extensions\ITaskItemExtensions.cs" LinkBase="Extensions" />
     <Compile Include="..\Mobile.BuildTools.Common\Logging\BuildHostLoggingHelper.cs" LinkBase="Logging" />
     <Compile Include="..\Mobile.BuildTools.Common\Tasks\BuildToolsTaskBase.cs" LinkBase="Tasks" />
   </ItemGroup>
