+++ src/Mobile.BuildTools.Images/MonoAndroid.targets
@@ -1,0 +1,75 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Project>
+
+  <UsingTask TaskName="Mobile.BuildTools.Tasks.GoogleTask"
+             AssemblyFile="$(_MobileBuildToolsParentBuildTasksDll)"/>
+
+  <Import Project="PlatformHead.targets" />
+  <Import Project="AndroidManifest.targets" />
+
+  <Target Name="_CollectAPK"
+          BeforeTargets="APKCopyToStagingDirectory">
+
+    <ItemGroup>
+      <__Generated_APK Include="$(OutputPath)\**\*.$(AndroidPackageFormat)" />
+    </ItemGroup>
+
+    <Message Text="@(__Generated_APK)" />
+
+  </Target>
+
+  <Target Name="APKCopyToStagingDirectory"
+          AfterTargets="_CreateBaseApk;_CollectAPK"
+          DependsOnTargets="_CreateBaseApk;_CollectAPK"
+          Condition=" '$(BuildToolsEnableArtifactCopy)' And !'$(IsAppCenter)' "
+          Inputs="@(__Generated_APK)"
+          Outputs="@(__Generated_APK -> '$(BuildToolsArtifactOutputPath)%(Filename)%(Extension)')">
+
+    <Copy SourceFiles="$(__Generated_APK)"
+          DestinationFolder="$(BuildToolsArtifactOutputPath)"
+          ContinueOnError="true"/>
+
+  </Target>
+
+  <PropertyGroup>
+    <HandlePlatformAppConfigBeforeTargets>_CollectBundleResources;PrepareForBuild</HandlePlatformAppConfigBeforeTargets>
+  </PropertyGroup>
+
+  <Target Name="HandlePlatformAppConfig"
+          AfterTargets="_CollectAppConfigs;ConfigurationManagerTransformAndCopy">
+    <ItemGroup>
+      <AndroidAsset Include="@(PlatformAppConfig)"
+                    Link="Assets\%(Filename)%(Extension)"/>
+    </ItemGroup>
+  </Target>
+
+  <Target Name="IncludeUnifiedImageAsset"
+          AfterTargets="UnifiedImageAssetProcessing"
+          BeforeTargets="UpdateAndroidResources"
+          Condition=" $(BuildToolsBundleUnifiedImages) == 'true' ">
+    <ItemGroup>
+      <AndroidResource Include="@(UnifiedImageAsset)"
+                       Link="%(UnifiedImageAsset.OutputLink)"
+                       Visible="%(UnifiedImageAsset.ShouldBeVisible)" />
+    </ItemGroup>
+  </Target>
+
+  <Target Name="MBTBundleGoogleServices"
+          BeforeTargets="Compile;CoreCompile">
+    <GoogleTask ConfigurationPath="$(BuildToolsConfigFilePath)"
+                ProjectName="$(MSBuildProjectName)"
+                ProjectDirectory="$(MSBuildProjectDirectory)"
+                SolutionDirectory="$(SolutionDir)"
+                Configuration="$(Configuration)"
+                IntermediateOutputPath="$(IntermediateOutputPath)"
+                TargetFrameworkIdentifier="$(TargetFrameworkIdentifier)">
+      <Output ItemName="GeneratedFile"
+              TaskParameter="GoogleOutput"/>
+      <Output ItemName="FilesWrite"
+              TaskParameter="GoogleOutput"/>
+      <Output ItemName="GoogleServicesJson"
+              TaskParameter="GoogleOutput" />
+    </GoogleTask>
+  </Target>
+
+</Project>
