+++ src/Mobile.BuildTools/Mobile.BuildTools.csproj
@@ -1,55 +1,41 @@
 <?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.Build.NoTargets/2.0.1">
 
   <PropertyGroup>
-    <TargetFramework>net472</TargetFramework>
+    <TargetFramework>netstandard1.0</TargetFramework>
     <PackageId>$(AssemblyName)</PackageId>
     <Description>The Mobile.BuildTools 2.0 is more powerful than ever before. Whether you're managing your application secrets, environment configurations, using CSS with Xamarin.Forms, or looking for a super powerful API that can enable powerful solutions for all of your App Icons / Images. If you aren't using Mobile.BuildTools with your Xamarin apps, you're working too hard or not following best practices! Truthfully I forgot everything it does so we wrote it down. Check out the new for 2.0 docs site at https://mobilebuildtools.com</Description>
     <PackageTags>msbuild;dotnet;xamarin;android;ios;macos;uwp;tizen;buildtools</PackageTags>
     <DevelopmentDependency>true</DevelopmentDependency>
     <IsPackable>true</IsPackable>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackFolder>build</PackFolder>
-    <IncludeSource>false</IncludeSource>
-    <PackageReadmeFile>readme.md</PackageReadmeFile>
+    <!--<PackageReadmeFile>readme.md</PackageReadmeFile>-->
+    <IsPackagingProject>true</IsPackagingProject>
   </PropertyGroup>
 
   <ItemGroup>
     <InternalsVisibleTo Include="Mobile.BuildTools.Tests" />
   </ItemGroup>
 
-  <ItemGroup>
+  <!--<ItemGroup>
     <None Include="*.targets;*.props" Pack="true" PackFolder="$(PackFolder)" />
-    <None Update="buildTransitive\*.targets;buildTransitive\*.props" Pack="true" PackFolder="buildTransitive" />
     <None Update="Resources\*" Pack="true" PackagePath="$(PackFolder)/resources/%(Filename)%(Extension)" />
     <None Update="readme.md" Pack="true" PackagePath="readme.md" />
-  </ItemGroup>
+  </ItemGroup>-->
 
   <ItemGroup>
-    <!--<PackageReference Include="HarfBuzzSharp" PrivateAssets="all" />-->
-    <PackageReference Include="Microsoft.Build" Pack="false" />
-    <PackageReference Include="Microsoft.Build.Tasks.Core" Pack="false" />
     <PackageReference Include="NuGetizer" PrivateAssets="all" />
-    <!--<PackageReference Include="System.Buffers" PrivateAssets="all" />
-    <PackageReference Include="System.IO.UnmanagedMemoryStream" PrivateAssets="all" />
-    <PackageReference Include="System.Numerics.Vectors" PrivateAssets="all" />
-    <PackageReference Include="System.Runtime.CompilerServices.Unsafe" PrivateAssets="all" />-->
-    <PackageReference Include="System.Text.Json" PrivateAssets="all" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Mobile.BuildTools.AppSettings\Mobile.BuildTools.AppSettings.csproj"
-                      ReferenceOutputAssembly="false" />
-    <ProjectReference Include="..\Mobile.BuildTools.FormsCss\Mobile.BuildTools.FormsCss.csproj"
-                      ReferenceOutputAssembly="false" />
-    <ProjectReference Include="..\Mobile.BuildTools.Images\Mobile.BuildTools.Images.csproj"
-                      ReferenceOutputAssembly="false" />
-    <ProjectReference Include="..\Mobile.BuildTools.Reference\Mobile.BuildTools.Reference.csproj" PackFolder="build" />
+    <ProjectReference Include="..\Mobile.BuildTools.AppPlatform\Mobile.BuildTools.AppPlatform.csproj" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\Mobile.BuildTools.AppSettings\Mobile.BuildTools.AppSettings.csproj" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\Mobile.BuildTools.FormsCss\Mobile.BuildTools.FormsCss.csproj" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\Mobile.BuildTools.Images\Mobile.BuildTools.Images.csproj" ReferenceOutputAssembly="false" />
+    <!--<ProjectReference Include="..\Mobile.BuildTools.Reference\Mobile.BuildTools.Reference.csproj" PackFolder="build" />-->
   </ItemGroup>
 
-  <Target Name="ExcludeTFM"
-          BeforeTargets="_AddPackageManifest"
-          AfterTargets="InferPackageContents">
+  <Target Name="ExcludeTFM" BeforeTargets="_AddPackageManifest" AfterTargets="InferPackageContents">
     <PropertyGroup>
       <IsPackagingProject>true</IsPackagingProject>
     </PropertyGroup>
