--- src/Mobile.BuildTools/buildTransitive/Mobile.BuildTools.props
@@ -1,3 +1,0 @@
-<Project>
-  <Import Project="$(MSBuildThisFileDirectory)..\build\Mobile.BuildTools.props" />
-</Project>
